+++ openBCIBoard.js
@@ -25,7 +25,9 @@ function OpenBCIFactory () {
     simulatorBoardFailure: false,
     simulatorDaisyModuleAttached: false,
     simulatorFirmwareVersion: k.OBCIFirmwareV1,
-    simulatorFragmentation: 'Random',
+    simulatorFragmentation: 'random',
+    simulatorLatencyTime: 16,
+    simulatorBufferSize: 4096,
     simulatorHasAccelerometer: true,
     simulatorInternalClockDrift: 0,
     simulatorInjectAlpha: true,
@@ -68,10 +70,16 @@ function OpenBCIFactory () {
   *              `v2` - Firmware Version 2
   *
   *     - `simulatorFragmentation` {String} - Specifies how to break packets to simulate fragmentation.
-  *          3 Possible Options:
-  *              `Random` - output randomly-lengthed chunks of data (Default)
-  *              `OneByOne` - output each byte separately
-  *              `None` - do not fragment packets; output complete chunks immediately after they are produced
+  *          4 Possible Options:
+  *              `random` - output random small chunks of data interspersed with full buffers (Default)
+  *              `fullBuffers` - allow buffers to fill up until the latency timer has expired
+  *              `oneByOne` - output each byte separately
+  *              `none` - do not fragment packets; output complete chunks immediately after they are produced
+  *
+  *     - `simulatorLatencyTime` {Number} - The time in milliseconds to wait before sending partially full buffers.
+  *                  (Default `16`)
+  *
+  *     - `simulatorBufferSize` {Number} - The size of a full buffer of data. (Default `4096`)
   *
   *     - `simulatorHasAccelerometer` - {Boolean} - Sets simulator to send packets with accelerometer data. (Default `true`)
   *
@@ -81,7 +89,7 @@ function OpenBCIFactory () {
   *          3 Possible Options:
   *              `60Hz` - 60Hz line noise (Default) [America]
   *              `50Hz` - 50Hz line noise [Europe]
-  *              `None` - Do not inject line noise.
+  *              `none` - Do not inject line noise.
   *
   *     - `simulatorSampleRate` {Number} - The sample rate to use for the simulator. Simulator will set to 125 if
   *                  `simulatorDaisyModuleAttached` is set `true`. However, setting this option overrides that
@@ -120,8 +128,8 @@ function OpenBCIFactory () {
       opts.simulatorFirmwareVersion = k.OBCIFirmwareV1;
     }
     opts.simulatorFragmentation = options.simulatorFragmentation || options.simulatorfragmentation || _options.simulatorFragmentation;
-    if (opts.simulatorFragmentation !== 'Random' && opts.simulatorFragmentation !== 'OneByOne' && opts.simulatorFragmentation !== 'None') {
-      opts.simulatorFragmentation = 'Random';
+    if (opts.simulatorFragmentation !== 'random' && opts.simulatorFragmentation !== 'fullBuffers' && opts.simulatorFragmentation !== 'oneByOne' && opts.simulatorFragmentation !== 'none') {
+      opts.simulatorFragmentation = 'random';
     }
     if (options.simulatorHasAccelerometer === false || options.simulatorhasaccelerometer === false) {
       opts.simulatorHasAccelerometer = false;
@@ -135,7 +143,7 @@ function OpenBCIFactory () {
       opts.simulatorInjectAlpha = _options.simulatorInjectAlpha;
     }
     opts.simulatorInjectLineNoise = options.simulatorInjectLineNoise || options.simulatorinjectlinenoise || _options.simulatorInjectLineNoise;
-    if (opts.simulatorInjectLineNoise !== '60Hz' && opts.simulatorInjectLineNoise !== '50Hz' && opts.simulatorInjectLineNoise !== 'None') {
+    if (opts.simulatorInjectLineNoise !== '60Hz' && opts.simulatorInjectLineNoise !== '50Hz' && opts.simulatorInjectLineNoise !== 'none') {
       opts.simulatorInjectLineNoise = '60Hz';
     }
     opts.simulatorSampleRate = options.simulatorSampleRate || options.simulatorsamplerate || _options.simulatorSampleRate;
@@ -245,6 +253,9 @@ function OpenBCIFactory () {
           daisy: this.options.simulatorDaisyModuleAttached,
           drift: this.options.simulatorInternalClockDrift,
           firmwareVersion: this.options.simulatorFirmwareVersion,
+          fragmentation: this.options.simulatorFragmentation,
+          latencyTime: this.options.simulatorLatencyTime,
+          bufferSize: this.options.simulatorBufferSize,
           lineNoise: this.options.simulatorInjectLineNoise,
           sampleRate: this.options.simulatorSampleRate,
           serialPortFailure: this.options.simulatorSerialPortFailure,